requirements.txt
click==7.0
tweepy==3.8.0
python-dotenv==0.10.3